Implement react router in the project
