+++ src/App.jsx
@@ -1,16 +1,9 @@
-import Heading from "./components/Heading";
-import Input from "./components/Input";
-import Select from "./components/Select";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
 
-import { useState } from "react";
+import CurrenciesListPage from "./pages/CurrenciesList";
+import DetailsPage from "./pages/Details";
 
 function App() {
-  const [inputValue, setInputValue] = useState("");
-  const handleInputChange = (event) => setInputValue(event.target.value);
-
-  const [selectValue, setSelectValue] = useState("");
-  const handleSelectChange = (event) => setSelectValue(event.target.value);
-
   const currencies = [
     { currency: "USD" },
     { currency: "PLN" },
@@ -19,20 +12,18 @@ function App() {
   ];
 
   return (
-    <>
-      <Heading variant="h1">Currency converter</Heading>
-      <Heading variant="h2">Choose base currency</Heading>
-      <Heading variant="h2">Convert values</Heading>
-      <Input value={inputValue} onChange={handleInputChange} />
-      <Select
-        value={selectValue}
-        onChange={handleSelectChange}
-        options={currencies.map(({ currency }) => ({
-          value: currency,
-          label: currency,
-        }))}
-      />
-    </>
+    <BrowserRouter>
+      <Routes>
+        <Route
+          path="/"
+          element={<CurrenciesListPage currencies={currencies} />}
+        />
+        <Route
+          path="/details"
+          element={<DetailsPage currencies={currencies} />}
+        />
+      </Routes>
+    </BrowserRouter>
   );
 }
